put util functions and type into different files
