+++ src/types/index.ts
@@ -17,3 +17,8 @@ export type ResponseObject = {
   code: number;
   meta?: unknown;
 };
+
+export const entityTypes = {
+	CLIENT: 'CLIENT',
+	BANK: 'BANK',
+} as const;
